requirements.txt
Pillow
zbarlight
mysqlclient
pytesseract